testKG/requirements.txt
flask>=2.0.1
neo4j>=4.4.3
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.5.0
numpy>=1.21.0 